[追加]コメントの最大取得数に関する設定を追加 (#136)
* [追加]デザイン用のVMを追加

* [変更]CommentCollectionクラスを別ファイルに移動

* [追加]コメントの最大取得数に関する設定を追加

* [修正]コメントにプレミアムフラグが付加されていなかった問題を修正
